LayoutTests/http/tests/model/model-document.html

<!doctype html>
<html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>

promise_test(async test => {
    const frame = document.createElement("iframe");
    frame.src = "UnitBox.usdz";
    document.body.appendChild(frame);

    await new Promise(resolve => frame.addEventListener("load", resolve, { once: true }));

    assert_true(!!frame.contentDocument.querySelector("model"), "The frame contains a <model> element");
    assert_true(!!frame.contentDocument.querySelector("model source"), "The <model> element contains a <source> element");
}, "Loading a standalone model resource creates a standalone <model> document.");

</script>
</body>
</html>
